Removed Duplicate singleton
