+++ UnityProject/Assets/Scripts/Managers/SingletonManager.cs
@@ -1,4 +1,3 @@
-using Mirror;
 using UnityEngine;
 
 namespace Managers
@@ -7,20 +6,35 @@ namespace Managers
 	/// Singleton Manager using static instances without use of FindObject
 	/// If you are using Awake() override and remember to call base.Awake()!
 	/// </summary>
-	public class SingletonManager<T> : MonoBehaviour where T : MonoBehaviour
+	public abstract class SingletonManager<T> : MonoBehaviour where T : MonoBehaviour
 	{
-		public static T Instance { get; private set; }
+		public static T Instance;
 
+
+		/// <summary>
+		/// If you override this then make sure you call base.Awake() somewhere in your Awake code.
+		/// </summary>
 		public virtual void Awake()
 		{
-			if (Instance == null)
+			Instance = this as T;
+		}
+
+
+		public virtual void Start()
+		{
+			Instance = this as T;
+		}
+
+		/// <summary>
+		/// If you override this then make sure you call base.OnDestroy() somewhere in your OnDestroy code.
+		/// </summary>
+		protected void OnDestroy()
+		{
+			if (Instance == this)
 			{
-				Instance = this as T;
-			}
-			else
-			{
-				Destroy(gameObject);
+				Instance = null;
 			}
 		}
 	}
+
 }
