--- UnityProject/Assets/Scripts/Util/ScriptsUtilMonoBehaviourSingleton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 16a2faa8729f81b4aa8d1959ade9e85a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
